+++ tests/adapter_root_spec.lua
@@ -14,7 +14,15 @@ describe("root when using solution option", function()
     local dir = "./tests/solution_dir"
     local root = plugin.root(dir)
 
-    assert.equal(dir, root)
+    assert.equal(vim.fs.normalize(vim.fn.fnamemodify(dir, ":p")), root)
+  end)
+
+  async.it("should return the .slnx dir when it exists and path contains it", function()
+    local plugin = require("neotest-dotnet")
+    local dir = "./tests/solutionx_dir"
+    local root = plugin.root(dir)
+
+    assert.equal(vim.fs.normalize(vim.fn.fnamemodify(dir, ":p")), root)
   end)
 
   async.it("should return nil when neither path nor parents contain .sln file", function()
@@ -51,7 +59,7 @@ describe("root when using project option", function()
     local dir = "./tests/project_dir"
     local root = plugin.root(dir)
 
-    assert.equal(dir, root)
+    assert.equal(vim.fs.normalize(vim.fn.fnamemodify(dir, ":p")), root)
   end)
 
   async.it("should return nil when neither path nor parents contain .csproj file", function()