+++ tests/nunit/specs/mixed_attributes.cs
@@ -1,0 +1,12 @@
+using NUnit.Framework;
+
+public class MixedAttributes
+{
+    [Test]
+    [TestCase(1)]
+    [TestCase(2)]
+    public void Mixed(int value)
+    {
+        Assert.That(value, Is.GreaterThan(0));
+    }
+}